+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add pod, container, node and node role labels
 - Allow ignoring clusters using the `giantswarm.io/monitoring: false` label on cluster CRs
 - Add monitoring of control plane bastions
+- Add heartbeat alert to prometheus
+- Create heartbeat in opsgenie
 
 ## [1.7.0] - 2020-10-14